Personal timesheet
Inspired by my partner's shift work and a wonder whether the payment received was accurate, I put together this workbook. This is my first complete Visual Basic project. The features of my timesheet are as follows:
- Enter in shifts on a monthly tab. This calculates the length of the shift (with or without an hour break) and the gross pay for that shift (with regard to the penalty rates).
- Shifts can be adjusted or deleted.
- At the end of the pay cycle (i.e. when the payslip is received), both gross and net pay can be calculated for that full period (currently set to 14 days).
- Each month is on a separate tab, and can be activated through the first tab's listbox.

This project is ongoing, there are a few matters which need to be expanded/resolved (as noted in the changelog).